add modules for EosVariant and FunctionalVariant so that they can be used outside of Python feature
