+++ src/python/eos.rs
@@ -1,3 +1,4 @@
+use crate::eos::EosVariant;
 #[cfg(feature = "estimator")]
 use crate::estimator::*;
 #[cfg(feature = "gc_pcsaft")]
@@ -24,7 +25,6 @@ use feos_core::cubic::PengRobinson;
 use feos_core::python::cubic::PyPengRobinsonParameters;
 use feos_core::python::user_defined::PyEoSObj;
 use feos_core::*;
-use ndarray::Array1;
 use numpy::convert::ToPyArray;
 use numpy::{PyArray1, PyArray2};
 use pyo3::exceptions::{PyIndexError, PyValueError};
@@ -35,181 +35,6 @@ use quantity::python::*;
 use quantity::si::*;
 use std::collections::HashMap;
 use std::rc::Rc;
-
-pub enum EosVariant {
-    #[cfg(feature = "pcsaft")]
-    PcSaft(PcSaft),
-    #[cfg(feature = "gc_pcsaft")]
-    GcPcSaft(GcPcSaft),
-    PengRobinson(PengRobinson),
-    Python(PyEoSObj),
-    #[cfg(feature = "pets")]
-    Pets(Pets),
-    #[cfg(feature = "uvtheory")]
-    UVTheory(UVTheory),
-}
-
-impl EquationOfState for EosVariant {
-    fn components(&self) -> usize {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.components(),
-            #[cfg(feature = "gc_pcsaft")]
-            EosVariant::GcPcSaft(eos) => eos.components(),
-            EosVariant::PengRobinson(eos) => eos.components(),
-            EosVariant::Python(eos) => eos.components(),
-            #[cfg(feature = "pets")]
-            EosVariant::Pets(eos) => eos.components(),
-            #[cfg(feature = "uvtheory")]
-            EosVariant::UVTheory(eos) => eos.components(),
-        }
-    }
-
-    fn compute_max_density(&self, moles: &Array1<f64>) -> f64 {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.compute_max_density(moles),
-            #[cfg(feature = "gc_pcsaft")]
-            EosVariant::GcPcSaft(eos) => eos.compute_max_density(moles),
-            EosVariant::PengRobinson(eos) => eos.compute_max_density(moles),
-            EosVariant::Python(eos) => eos.compute_max_density(moles),
-            #[cfg(feature = "pets")]
-            EosVariant::Pets(eos) => eos.compute_max_density(moles),
-            #[cfg(feature = "uvtheory")]
-            EosVariant::UVTheory(eos) => eos.compute_max_density(moles),
-        }
-    }
-
-    fn subset(&self, component_list: &[usize]) -> Self {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => Self::PcSaft(eos.subset(component_list)),
-            #[cfg(feature = "gc_pcsaft")]
-            EosVariant::GcPcSaft(eos) => Self::GcPcSaft(eos.subset(component_list)),
-            EosVariant::PengRobinson(eos) => Self::PengRobinson(eos.subset(component_list)),
-            EosVariant::Python(eos) => Self::Python(eos.subset(component_list)),
-            #[cfg(feature = "pets")]
-            EosVariant::Pets(eos) => Self::Pets(eos.subset(component_list)),
-            #[cfg(feature = "uvtheory")]
-            EosVariant::UVTheory(eos) => Self::UVTheory(eos.subset(component_list)),
-        }
-    }
-
-    fn residual(&self) -> &[Box<dyn HelmholtzEnergy>] {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.residual(),
-            #[cfg(feature = "gc_pcsaft")]
-            EosVariant::GcPcSaft(eos) => eos.residual(),
-            EosVariant::PengRobinson(eos) => eos.residual(),
-            EosVariant::Python(eos) => eos.residual(),
-            #[cfg(feature = "pets")]
-            EosVariant::Pets(eos) => eos.residual(),
-            #[cfg(feature = "uvtheory")]
-            EosVariant::UVTheory(eos) => eos.residual(),
-        }
-    }
-
-    fn ideal_gas(&self) -> &dyn IdealGasContribution {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.ideal_gas(),
-            #[cfg(feature = "gc_pcsaft")]
-            EosVariant::GcPcSaft(eos) => eos.ideal_gas(),
-            EosVariant::PengRobinson(eos) => eos.ideal_gas(),
-            EosVariant::Python(eos) => eos.ideal_gas(),
-            #[cfg(feature = "pets")]
-            EosVariant::Pets(eos) => eos.ideal_gas(),
-            #[cfg(feature = "uvtheory")]
-            EosVariant::UVTheory(eos) => eos.ideal_gas(),
-        }
-    }
-}
-
-impl MolarWeight<SIUnit> for EosVariant {
-    fn molar_weight(&self) -> SIArray1 {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.molar_weight(),
-            #[cfg(feature = "gc_pcsaft")]
-            EosVariant::GcPcSaft(eos) => eos.molar_weight(),
-            EosVariant::PengRobinson(eos) => eos.molar_weight(),
-            EosVariant::Python(eos) => eos.molar_weight(),
-            #[cfg(feature = "pets")]
-            EosVariant::Pets(eos) => eos.molar_weight(),
-            #[cfg(feature = "uvtheory")]
-            EosVariant::UVTheory(_) => unimplemented!(),
-        }
-    }
-}
-
-#[cfg(feature = "pcsaft")]
-impl EntropyScaling<SIUnit> for EosVariant {
-    fn viscosity_reference(
-        &self,
-        temperature: SINumber,
-        volume: SINumber,
-        moles: &SIArray1,
-    ) -> EosResult<SINumber> {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.viscosity_reference(temperature, volume, moles),
-            _ => unimplemented!(),
-        }
-    }
-
-    fn viscosity_correlation(&self, s_res: f64, x: &Array1<f64>) -> EosResult<f64> {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.viscosity_correlation(s_res, x),
-            _ => unimplemented!(),
-        }
-    }
-
-    fn diffusion_reference(
-        &self,
-        temperature: SINumber,
-        volume: SINumber,
-        moles: &SIArray1,
-    ) -> EosResult<SINumber> {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.diffusion_reference(temperature, volume, moles),
-            _ => unimplemented!(),
-        }
-    }
-
-    fn diffusion_correlation(&self, s_res: f64, x: &Array1<f64>) -> EosResult<f64> {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.diffusion_correlation(s_res, x),
-            _ => unimplemented!(),
-        }
-    }
-
-    fn thermal_conductivity_reference(
-        &self,
-        temperature: SINumber,
-        volume: SINumber,
-        moles: &SIArray1,
-    ) -> EosResult<SINumber> {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => {
-                eos.thermal_conductivity_reference(temperature, volume, moles)
-            }
-            _ => unimplemented!(),
-        }
-    }
-
-    fn thermal_conductivity_correlation(&self, s_res: f64, x: &Array1<f64>) -> EosResult<f64> {
-        match self {
-            #[cfg(feature = "pcsaft")]
-            EosVariant::PcSaft(eos) => eos.thermal_conductivity_correlation(s_res, x),
-            _ => unimplemented!(),
-        }
-    }
-}
 
 #[pyclass(name = "EquationOfState", unsendable)]
 #[derive(Clone)]
